Vkhydzik lvneta211 apply the filter to the required controllers and fix the unit tests for the controller and service for refresh token (#102)
* Fixes the unit tests

* Apply the filter to the required controllers

+++ TheaterSchedule.BLL/Interfaces/IRefreshTokenService.cs
@@ -6,10 +6,10 @@ namespace TheaterSchedule.BLL.Interfaces
 {
     public interface IRefreshTokenService
     {
-        Task AddRefreshTokenAsync(string refreshToken, int userId, double daysToExpire);
+        Task AddRefreshTokenAsync(string refreshToken, int userId, int daysToExpire);
         string GenerateRefreshToken(int size = 32);
         Task<RefreshTokenDTO> GetAsync(string refreshToken);
         Task<TokensResponse> CheckRefreshTokenAsync(string refreshToken);
-        Task UpdateRefreshTokenAsync(int id, string refreshToken, int userId, double daysToExpire, bool isActive = true);
+        Task UpdateRefreshTokenAsync(int id, string refreshToken, int userId, int daysToExpire, bool isActive = true);
     }
 }
